Add files via upload
This modification to Mr. Blinky's code allows the user to use an MPU6050 accelerometer for the left and right direectional inputs. In order to run, the Arduboy2 core needs to be modified to enable the serial communication port on the ATMega tiny. To do so  remove _BV(PRTWI) from the Arduboy2Core::bootPowerSavings() function. Connect the SCL, SDA, 3.3V and GND pins of the Arduino Micro and the MPU6050.